Error checking + unit test added.

+++ common/main/src/uk/ac/ox/cs/pdq/databasemanagement/sqlcommands/Delete.java
@@ -23,7 +23,9 @@ public class Delete extends Command {
 	 * @param schema
 	 *            - schema is needed to get the relation (the fact contains a
 	 *            predicate only that might not be a relation)
-	 * @throws DatabaseException in case the input fact contains some error. For example it is not connected to the given schema by having an unknown predicate.
+	 * @throws DatabaseException
+	 *             in case the input fact contains some error. For example it is not
+	 *             connected to the given schema by having an unknown predicate.
 	 */
 	public Delete(Atom factToDelete, Schema schema) throws DatabaseException {
 		super();
@@ -40,14 +42,22 @@ public class Delete extends Command {
 	 * Creates a delete form statement string
 	 * 
 	 * @return
-	 * @throws DatabaseException 
+	 * @throws DatabaseException
 	 */
 	protected String createDeleteStatement(Atom fact, Schema schema) throws DatabaseException {
+		if (fact == null || schema == null)
+			throw new DatabaseException("Cant delete unset fact or from an unset schema. Fact: " + fact + ", schema: " + schema);
 		// header
 		String deleteFrom = "DELETE FROM " + DATABASENAME + "." + fact.getPredicate().getName() + " WHERE ";
 		Relation r = schema.getRelation(fact.getPredicate().getName());
-		int index = 0;
 
+		if (r == null)
+			throw new DatabaseException("Fact : " + fact + " doesn't belong to schema " + schema);
+
+		if (r.getAttributes().length != fact.getTerms().length)
+			throw new DatabaseException("Fact have different number of terms then the attributes of the relation: " + fact + ", relation " + r);
+
+		int index = 0;
 		// condition
 		for (Attribute a : r.getAttributes()) {
 			if (index != 0)
